pyproject.toml
[project]
name = "lumimind"
version = "0.1.0"
description = "Control your lamps in natural language"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dashscope>=1.22.2",
    "funasr>=1.2.6",
    "loguru>=0.7.3",
    "openai>=1.68.2",
    "opuslib>=3.0.1",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.0",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.2",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "I", # isort
    "E", # pycodestyle error
    "W", # pycodestyle warning
    "F", # Pyflakes
    "B", # flake8-bugbear
]
